+++ src/main/resources/jeecg/ext-template/nopop/single/cgform_controllerTemplate.ftl
@@ -13,7 +13,8 @@ import java.text.SimpleDateFormat;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
@@ -50,7 +51,7 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 import java.util.Map;
 import org.jeecgframework.core.util.ExceptionUtil;
-
+<#if cgformConfig.supportRestful?? && cgformConfig.supportRestful == "1">
 <#-- restful 通用方法生成 -->
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -69,12 +70,26 @@ import javax.validation.Validator;
 import java.net.URI;
 import org.springframework.http.MediaType;
 import org.springframework.web.util.UriComponentsBuilder;
+import org.apache.commons.lang3.StringUtils;
+import org.jeecgframework.jwt.util.GsonUtil;
+import org.jeecgframework.jwt.util.ResponseMessage;
+import org.jeecgframework.jwt.util.Result;
+import com.alibaba.fastjson.JSONArray;
 <#-- restful 通用方法生成 -->
 
+<#-- swagger api  start -->
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+<#-- swagger api end -->
+</#if>
 <#-- 列为文件类型的文件代码生成 -->
 <#assign fileFlag = false />
 <#list columns as filePo>
-	<#if filePo.showType=='file'>
+<#-- update--begin--author:gj_shaojc date:20180302 for：TASK #2551 【bug】网友问题验证确认 -->
+	<#if filePo.showType=='file'  || filePo.showType == 'image'>
+<#-- update--end--author:gj_shaojc date:20180302 for：TASK #2551 【bug】网友问题验证确认 -->
 		<#assign fileFlag = true />
 	</#if>
 </#list>
@@ -93,21 +108,25 @@ import java.util.HashMap;
  * @version V1.0   
  *
  */
+ <#if cgformConfig.supportRestful?? && cgformConfig.supportRestful == "1">
+ <#-- update--begin--author:zhangjiaqiang date:20171031 for:API 注解 start -->
+@Api(value="${entityName}",description="${ftl_description}",tags="${entityName?uncap_first}Controller")
+<#-- update--end--author:zhangjiaqiang date:20171031 for:API 注解 start -->
+</#if>
 @Controller
 @RequestMapping("/${entityName?uncap_first}Controller")
 public class ${entityName}Controller extends BaseController {
-	/**
-	 * Logger for this class
-	 */
-	private static final Logger logger = Logger.getLogger(${entityName}Controller.class);
+	private static final Logger logger = LoggerFactory.getLogger(${entityName}Controller.class);
 
 	@Autowired
 	private ${entityName}ServiceI ${entityName?uncap_first}Service;
 	@Autowired
 	private SystemService systemService;
 	<#-- restful 通用方法生成 -->
+	<#if cgformConfig.supportRestful?? && cgformConfig.supportRestful == "1">
 	@Autowired
 	private Validator validator;
+	</#if>
 	<#-- restful 通用方法生成 -->
 	
 	<#-- 列为文件类型的文件代码生成 -->
@@ -145,26 +164,6 @@ public class ${entityName}Controller extends BaseController {
 		org.jeecgframework.core.extend.hqlsearch.HqlGenerateUtil.installHql(cq, ${entityName?uncap_first}, request.getParameterMap());
 		try{
 		//自定义追加查询条件
-		<#list columns as po>
-		<#if po.isQuery =='Y' && po.queryMode =='group'>
-		String query_${po.fieldName}_begin = request.getParameter("${po.fieldName}_begin");
-		String query_${po.fieldName}_end = request.getParameter("${po.fieldName}_end");
-		if(StringUtil.isNotEmpty(query_${po.fieldName}_begin)){
-			<#if po.type == "java.util.Date">
-			cq.ge("${po.fieldName}", new SimpleDateFormat("yyyy-MM-dd").parse(query_${po.fieldName}_begin));
-			<#else>
-			cq.ge("${po.fieldName}", Integer.parseInt(query_${po.fieldName}_begin));
-			</#if>
-		}
-		if(StringUtil.isNotEmpty(query_${po.fieldName}_end)){
-			<#if po.type == "java.util.Date">
-			cq.le("${po.fieldName}", new SimpleDateFormat("yyyy-MM-dd").parse(query_${po.fieldName}_end));
-			<#else>
-			cq.le("${po.fieldName}", Integer.parseInt(query_${po.fieldName}_end));
-			</#if>
-		}
-		</#if>
-		</#list> 
 		}catch (Exception e) {
 			throw new BusinessException(e.getMessage());
 		}
@@ -368,7 +367,7 @@ public class ${entityName}Controller extends BaseController {
 		List<${entityName}Entity> ${entityName?uncap_first}s = this.${entityName?uncap_first}Service.getListByCriteriaQuery(cq,false);
 		modelMap.put(NormalExcelConstants.FILE_NAME,"${ftl_description}");
 		modelMap.put(NormalExcelConstants.CLASS,${entityName}Entity.class);
-		modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("${ftl_description}列表", "导出人:"+ResourceUtil.getSessionUserName().getRealName(),
+		modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("${ftl_description}列表", "导出人:"+ResourceUtil.getSessionUser().getRealName(),
 			"导出信息"));
 		modelMap.put(NormalExcelConstants.DATA_LIST,${entityName?uncap_first}s);
 		return NormalExcelConstants.JEECG_EXCEL_VIEW;
@@ -384,7 +383,7 @@ public class ${entityName}Controller extends BaseController {
 			, DataGrid dataGrid,ModelMap modelMap) {
     	modelMap.put(NormalExcelConstants.FILE_NAME,"${ftl_description}");
     	modelMap.put(NormalExcelConstants.CLASS,${entityName}Entity.class);
-    	modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("${ftl_description}列表", "导出人:"+ResourceUtil.getSessionUserName().getRealName(),
+    	modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("${ftl_description}列表", "导出人:"+ResourceUtil.getSessionUser().getRealName(),
     	"导出信息"));
     	modelMap.put(NormalExcelConstants.DATA_LIST,new ArrayList());
     	return NormalExcelConstants.JEECG_EXCEL_VIEW;
@@ -412,7 +411,7 @@ public class ${entityName}Controller extends BaseController {
 				j.setMsg("文件导入成功！");
 			} catch (Exception e) {
 				j.setMsg("文件导入失败！");
-				logger.error(ExceptionUtil.getExceptionMessage(e));
+				logger.error(e.getMessage());
 			}finally{
 				try {
 					file.getInputStream().close();
@@ -455,31 +454,50 @@ public class ${entityName}Controller extends BaseController {
 	</#if>
 	<#-- 列为文件类型的文件代码生成 -->
 	
+	<#if cgformConfig.supportRestful?? && cgformConfig.supportRestful == "1">
+	<#-- update--begin--author:zhangjiaqiang date:20171113 for:restful接口封装 -->
 	<#-- restful 通用方法生成 -->
-	@RequestMapping(method = RequestMethod.GET)
+	<#-- update-begin-Author:LiShaoQing Date:20180828 for: TASK #3105 【代码生成器】代码生成rest接口 list获取改造 -->
+	@RequestMapping(value="/list/{pageNo}/{pageSize}", method = RequestMethod.GET)
 	@ResponseBody
-	public List<${entityName}Entity> list() {
-		List<${entityName}Entity> list${entityName}s=${entityName?uncap_first}Service.getList(${entityName}Entity.class);
-		return list${entityName}s;
+	<#-- update--begin--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	@ApiOperation(value="${ftl_description}列表信息",produces="application/json",httpMethod="GET")
+	<#-- update--end--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	public ResponseMessage<List<${entityName}Entity>> list(@PathVariable("pageNo") int pageNo, @PathVariable("pageSize") int pageSize, HttpServletRequest request) {
+		if(pageSize > Globals.MAX_PAGESIZE){
+			return Result.error("每页请求不能超过" + Globals.MAX_PAGESIZE + "条");
+		}
+		CriteriaQuery query = new CriteriaQuery(${entityName}Entity.class);
+		query.setCurPage(pageNo<=0?1:pageNo);
+		query.setPageSize(pageSize<1?1:pageSize);
+		List<${entityName}Entity> list${entityName}s = this.${entityName?uncap_first}Service.getListByCriteriaQuery(query,true);
+		return Result.success(list${entityName}s);
 	}
+	<#-- update-end-Author:LiShaoQing Date:20180828 for: TASK #3105 【代码生成器】代码生成rest接口 list获取改造 -->
 	
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
 	@ResponseBody
-	public ResponseEntity<?> get(@PathVariable("id") String id) {
+	<#-- update--begin--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	@ApiOperation(value="根据ID获取${ftl_description}信息",notes="根据ID获取${ftl_description}信息",httpMethod="GET",produces="application/json")
+	<#-- update--end--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	public ResponseMessage<?> get(@ApiParam(required=true,name="id",value="ID")@PathVariable("id") String id) {
 		${entityName}Entity task = ${entityName?uncap_first}Service.get(${entityName}Entity.class, id);
 		if (task == null) {
-			return new ResponseEntity(HttpStatus.NOT_FOUND);
+			return Result.error("根据ID获取${ftl_description}信息为空");
 		}
-		return new ResponseEntity(task, HttpStatus.OK);
+		return Result.success(task);
 	}
 
 	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
 	@ResponseBody
-	public ResponseEntity<?> create(@RequestBody ${entityName}Entity ${entityName?uncap_first}, UriComponentsBuilder uriBuilder) {
+	<#-- update--begin--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	@ApiOperation(value="创建${ftl_description}")
+	public ResponseMessage<?> create(@ApiParam(name="${ftl_description}对象")@RequestBody ${entityName}Entity ${entityName?uncap_first}, UriComponentsBuilder uriBuilder) {
+		<#-- update--end--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
 		//调用JSR303 Bean Validator进行校验，如果出错返回含400错误码及json格式的错误信息.
 		Set<ConstraintViolation<${entityName}Entity>> failures = validator.validate(${entityName?uncap_first});
 		if (!failures.isEmpty()) {
-			return new ResponseEntity(BeanValidators.extractPropertyAndMessage(failures), HttpStatus.BAD_REQUEST);
+			return Result.error(JSONArray.toJSONString(BeanValidators.extractPropertyAndMessage(failures)));
 		}
 
 		//保存
@@ -487,23 +505,21 @@ public class ${entityName}Controller extends BaseController {
 			${entityName?uncap_first}Service.save(${entityName?uncap_first});
 		} catch (Exception e) {
 			e.printStackTrace();
-			return new ResponseEntity(HttpStatus.NO_CONTENT);
+			return Result.error("${ftl_description}信息保存失败");
 		}
-		//按照Restful风格约定，创建指向新任务的url, 也可以直接返回id或对象.
-		String id = ${entityName?uncap_first}.getId();
-		URI uri = uriBuilder.path("/rest/${entityName?uncap_first}Controller/" + id).build().toUri();
-		HttpHeaders headers = new HttpHeaders();
-		headers.setLocation(uri);
-
-		return new ResponseEntity(headers, HttpStatus.CREATED);
+		return Result.success(${entityName?uncap_first});
 	}
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<?> update(@RequestBody ${entityName}Entity ${entityName?uncap_first}) {
+	@ResponseBody
+	<#-- update--begin--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	@ApiOperation(value="更新${ftl_description}",notes="更新${ftl_description}")
+	<#-- update--end--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	public ResponseMessage<?> update(@ApiParam(name="${ftl_description}对象")@RequestBody ${entityName}Entity ${entityName?uncap_first}) {
 		//调用JSR303 Bean Validator进行校验，如果出错返回含400错误码及json格式的错误信息.
 		Set<ConstraintViolation<${entityName}Entity>> failures = validator.validate(${entityName?uncap_first});
 		if (!failures.isEmpty()) {
-			return new ResponseEntity(BeanValidators.extractPropertyAndMessage(failures), HttpStatus.BAD_REQUEST);
+			return Result.error(JSONArray.toJSONString(BeanValidators.extractPropertyAndMessage(failures)));
 		}
 
 		//保存
@@ -511,17 +527,34 @@ public class ${entityName}Controller extends BaseController {
 			${entityName?uncap_first}Service.saveOrUpdate(${entityName?uncap_first});
 		} catch (Exception e) {
 			e.printStackTrace();
-			return new ResponseEntity(HttpStatus.NO_CONTENT);
+			return Result.error("更新${ftl_description}信息失败");
 		}
 
 		//按Restful约定，返回204状态码, 无内容. 也可以返回200状态码.
-		return new ResponseEntity(HttpStatus.NO_CONTENT);
+		return Result.success("更新${ftl_description}信息成功");
 	}
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable("id") String id) {
-		${entityName?uncap_first}Service.deleteEntityById(${entityName}Entity.class, id);
+	<#-- update--begin--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	@ApiOperation(value="删除${ftl_description}")
+	<#-- update--begin--author:zhangjiaqiang date:20171031 for:TASK #2397 【新功能】代码生成器模板修改，追加swagger-ui注解 -->
+	public ResponseMessage<?> delete(@ApiParam(name="id",value="ID",required=true)@PathVariable("id") String id) {
+		logger.info("delete[{}]" , id);
+		// 验证
+		if (StringUtils.isEmpty(id)) {
+			return Result.error("ID不能为空");
+		}
+		try {
+			${entityName?uncap_first}Service.deleteEntityById(${entityName}Entity.class, id);
+		} catch (Exception e) {
+			e.printStackTrace();
+			return Result.error("${ftl_description}删除失败");
+		}
+
+		return Result.success();
 	}
 	<#-- restful 通用方法生成 -->
+	<#-- update--end--author:zhangjiaqiang date:20171113 for:restful接口封装 -->
+	</#if>
 }
